--- app/src/main/java/com/musinsa/app/AppApplication.java
@@ -1,13 +1,0 @@
-package com.musinsa.app;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class AppApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(AppApplication.class, args);
-    }
-
-}
