--- app/src/main/java/com/musinsa/app/product/controller/ProductController.java
@@ -1,4 +1,0 @@
-package com.musinsa.app.product.controller;
-
-public class ProductController {
-}
