--- app/src/main/java/com/musinsa/app/product/model/dto/ProductDTO.java
@@ -1,4 +1,0 @@
-package com.musinsa.app.product.model.dto;
-
-public class ProductDTO {
-}
