--- app/src/main/java/com/musinsa/app/product/model/entity/Category.java
@@ -1,26 +1,0 @@
-package com.musinsa.app.product.model.entity;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@Entity
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "category")
-public class Category {
-
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
-
-	@Column(nullable = false, unique = true)
-	private String name;
-
-	@Builder
-	public Category( String name) {
-		this.name = name;
-	}
-}
