--- app/src/main/java/com/musinsa/app/product/model/entity/Product.java
@@ -1,35 +1,0 @@
-package com.musinsa.app.product.model.entity;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@Entity
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "products")
-public class Product {
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
-
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "category_id", nullable = false)
-	private Category category;
-
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "brand_id", nullable = false)
-	private Brand brand;
-
-	@Column(nullable = false)
-	private double price;
-
-	@Builder
-	protected Product(double price, Category category, Brand brand) {
-		this.price = price;
-		this.category = category;
-		this.brand = brand;
-	}
-}
