--- app/src/main/java/com/musinsa/app/product/repository/ProductRepository.java
@@ -1,4 +1,0 @@
-package com.musinsa.app.product.repository;
-
-public class ProductRepository {
-}
