--- app/src/main/java/com/musinsa/app/product/service/ProductService.java
@@ -1,4 +1,0 @@
-package com.musinsa.app.product.service;
-
-public class ProductService {
-}
