--- app/src/main/resources/application.properties
@@ -1,17 +1,0 @@
-spring.application.name=app
-
-spring.datasource.url= jdbc:h2:tcp://localhost/~/test
-spring.datasource.driverClassName=org.h2.Driver
-spring.datasource.username=chung
-spring.datasource.password=musinsa119
-spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
-spring.jpa.hibernate.ddl-auto=create
-spring.h2.console.enabled=true
-spring.jpa.defer-datasource-initialization=true
-spring.sql.init.mode=always
-spring.jpa.show-sql=true
-spring.jpa.properties.hibernate.format_sql=true
-spring.jpa.properties.hibernate.use_sql_comments=true
-
-logging.level.org.hibernate.SQL=DEBUG
-logging.level.org.hibernate.type.descriptor.sql.BasicBinder=TRACE
