--- app/src/test/java/com/musinsa/app/AppApplicationTests.java
@@ -1,13 +1,0 @@
-package com.musinsa.app;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class AppApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
