+++ product-app/src/main/java/com/musinsa/productapp/ProductAppApplication.java
@@ -1,0 +1,13 @@
+package com.musinsa.productapp;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ProductAppApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(ProductAppApplication.class, args);
+    }
+
+}
