+++ product-app/src/main/java/com/musinsa/productapp/ServletInitializer.java
@@ -1,0 +1,13 @@
+package com.musinsa.productapp;
+
+import org.springframework.boot.builder.SpringApplicationBuilder;
+import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
+
+public class ServletInitializer extends SpringBootServletInitializer {
+
+    @Override
+    protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
+        return application.sources(ProductAppApplication.class);
+    }
+
+}
