+++ product-app/src/main/java/com/musinsa/productapp/product/controller/ProductController.java
@@ -1,0 +1,4 @@
+package com.musinsa.productapp.product.controller;
+
+public class ProductController {
+}
