+++ product-app/src/main/java/com/musinsa/productapp/product/model/dto/ProductDTO.java
@@ -1,0 +1,4 @@
+package com.musinsa.productapp.product.model.dto;
+
+public class ProductDTO {
+}
