+++ product-app/src/main/java/com/musinsa/productapp/product/repository/ProductRepository.java
@@ -1,0 +1,4 @@
+package com.musinsa.productapp.product.repository;
+
+public class ProductRepository {
+}
