+++ product-app/src/main/java/com/musinsa/productapp/product/service/ProductService.java
@@ -1,0 +1,4 @@
+package com.musinsa.productapp.product.service;
+
+public class ProductService {
+}
