+++ product-app/src/test/java/com/musinsa/productapp/ProductAppApplicationTests.java
@@ -1,0 +1,13 @@
+package com.musinsa.productapp;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ProductAppApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
